models::krate: Inline `encodable()` method
